Added z axis momentum check to mulligan activation
The check to see if a player could use a mulligan only checked if
player.speed was zero, which does not account for vertical momentum. If
the player was in a situation where they only had veritcal momentum they
could not reset their shot. In particular this mean the player could get
softlocked by getting stuck on a vertical spring with no forward
momentum. This can now be escaped from.