Set spindash flag to true when starting a shot
I am hoping this will allow activation of the spindash door switch in
Deep Sea Zone Act 2. I have not tested that yet but it has had the nice
side effect of activating the spindash animation while charging a shot
and having Snolf emit the usual smoke particles for it. I was worried
this would result in unwanted behaviour or the player actually
performing a full spindash whenever they tried to do a shot, but it
doesn't seem to have had any immediate obvious negative consequences.