Simplify mulligan condition to only check if player can take a shot
There were cases where a player could be stuck in a state where they
were not allowed to take a shot (player.snolf.state = 3) but they did
also not meet the conditions for taking a mulligan, leaving them stuck
in place forever. Particilary this happened while rolling down a hill
into a corner. Changed the mulligan condition to simply be allowed if
the player is in state 3.