Remove quick turn ability.
The quick turn ability was only added because originally the player
could not change direction in 2D sections otherwise. At some point
changes that were made seem to have fixed that. That was probably just
when I stopped eating the controls and set acceleration to 0 instead and
I never retested 2D sections after that. As it stands it's unneeded and
just interferes with controls for characters that use that button with
Everybody's Snolf.