Only force roll state when player is on the ground
Don't force roll state every frame any more, only do it if the player is
on the ground. But set roll state when a shot is taken. This means
players will be put into the roll animation at the start of a shot but
can cancel into normal animations for air abilties.